Capstone_Team_Project/requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bitstring==3.1.9
bleach==4.1.0
cached-property==1.5.2
cffi==1.15.0
cycler==0.10.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
enum-compat==0.0.3
gast==0.5.2
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
muselsl==2.1.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg-resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.1
ptyprocess==0.7.0
pycparser==2.20
pygatt==4.0.5
Pygments==2.10.0
pylsl==1.10.5
pyparsing==3.0.1
pyrsistent==0.18.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.3
zipp==3.6.0
